# perf: improve constructing dist table