@BubbleCal 2024-11-22
dot distance type doesn't benefit much from this optimization, any ideas? @eddyxu