@eddyxu 2024-11-22
>dot distance type doesn't benefit much from this optimization, any ideas?
do we know if this is memory-bandwidth bound or just compute bound? we can look into that later.