@BubbleCal 2024-11-23
> >dot distance type doesn't benefit much from this optimization, any ideas?
> 
> 
> 
> do we know if this is memory-bandwidth bound or just compute bound? we can look into that later.

I think it's still compute bound, because L2 is much faster than dot, I'm thinking whether the 1-dot way leads to the compiler can't generate good enough code